Feat: 대댓글; 좋아요 CRUD 추가

+++ src/main/java/com/sparta/moviecomunnity/service/RecommentService.java
@@ -1,6 +1,7 @@
 package com.sparta.moviecomunnity.service;
 
 import com.sparta.moviecomunnity.dto.RecommentRequestDto;
+import com.sparta.moviecomunnity.dto.RecommentResponseDto;
 import com.sparta.moviecomunnity.entity.*;
 import com.sparta.moviecomunnity.exception.CustomException;
 import com.sparta.moviecomunnity.exception.ServerResponse;
@@ -10,10 +11,12 @@ import com.sparta.moviecomunnity.repository.RecommentRepository;
 import com.sparta.moviecomunnity.repository.UserRepository;
 import com.sparta.moviecomunnity.security.UserDetailsImpl;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -21,17 +24,44 @@ import static com.sparta.moviecomunnity.exception.ResponseCode.*;
 
 @RequiredArgsConstructor
 @Service
+@Slf4j
 public class RecommentService {
     private final UserRepository userRepository;
     private final CommentRepository commentRepository;
     private final RecommentRepository recommentRepository;
     private final JwtUtil jwtUtil;
 
-    // 대댓글 모아보기
-    public List<Recomment> getRecomments(String username) {
-        Optional<User> user = userRepository.findByUsername(username);
-        List<Recomment> recomments = recommentRepository.findAllByUser(user.get());
-        return recomments;
+    // 내가 쓴 대댓글 모아보기
+    @Transactional(readOnly = true)
+    public List<RecommentResponseDto> getRecomments(User user) {
+        List<RecommentResponseDto> toRecomments = new ArrayList<>();
+
+        if(user.getRole() == UserRoleEnum.ADMIN) {
+            List<Recomment> allRecomments = recommentRepository.findAll();
+            if(allRecomments != null) {
+                allRecomments.stream().forEach(
+                        (recomment) -> toRecomments.add(
+                                new RecommentResponseDto(recomment.getId(), recomment.getContext(),
+                                        recomment.getUser().getId(), recomment.getComment().getId(), recomment.getHearts().size())
+                        )
+                );
+            }
+        }
+
+//        유저일때
+        else {
+            List<Recomment> recomments = recommentRepository.findAllByUser(user);
+            if(recomments != null)  {
+                recomments.stream().forEach(
+                        (recomment) -> toRecomments.add(
+                                new RecommentResponseDto(recomment.getId(), recomment.getContext(),
+                                        recomment.getUser().getId(), recomment.getComment().getId(), recomment.getHearts().size())
+                        )
+                );
+            }
+        }
+
+        return toRecomments;
     }
 
     //대댓글 작성
@@ -52,21 +82,38 @@ public class RecommentService {
 
     //대댓글 수정
     @Transactional
-    public ResponseEntity<ServerResponse> editRecomment(Long id, RecommentRequestDto requestDto) {
+    public ResponseEntity<ServerResponse> editRecomment(Long id, RecommentRequestDto requestDto, User user) {
         Recomment recomment = recommentRepository.findRecommentById(id).orElseThrow(
                 () -> new CustomException(RESOURCE_NOT_FOUND)
         );
-        recomment.rewrite(requestDto.getContext());
-        recommentRepository.save(recomment);
+
+        List<Recomment> recomments = recommentRepository.findAllByUser(user);
+        if(recomments == null) {
+            throw new CustomException(INVALID_AUTH_TOKEN);
+        }
+
+        String newContext = requestDto.getContext();
+
+        if ( newContext != null || !newContext.isEmpty() ) {
+            recomment.rewrite(requestDto.getContext());
+            recommentRepository.save(recomment);
+        }
+
         return ServerResponse.toResponseEntity(SUCCESS_EDIT);
     }
 
     //대댓글 삭제
     @Transactional
-    public ResponseEntity<ServerResponse> deleteRecomment(long id) {
+    public ResponseEntity<ServerResponse> deleteRecomment(long id, User user) {
         Recomment recomment = recommentRepository.findById(id).orElseThrow(
                 () -> new CustomException(RESOURCE_NOT_FOUND)
         );
+
+        if(user.getRole() != UserRoleEnum.ADMIN && !recomment.getUser().getUsername().equals(user.getUsername())) {
+            throw new CustomException((INVALID_AUTH_TOKEN));
+        }
+
+        log.info("delete : {}", recomment.toString());
         recommentRepository.deleteById(id);
         return ServerResponse.toResponseEntity(SUCCESS_DELETE);
     }
